Consolidate `@testing-library/jest-dom`
