+++ apps/synapse-portal-framework/src/components/eliteportal/ELContributeYourData.test.tsx
@@ -1,5 +1,4 @@
 import { render, screen } from '@testing-library/react'
-import '@testing-library/jest-dom'
 import ELContributeYourData from './ELContributeYourData'
 import { describe, expect, test } from 'vitest'
