+++ apps/synapse-portal-framework/tsconfig.json
@@ -7,7 +7,7 @@
     "isolatedModules": true,
     "noFallthroughCasesInSwitch": true,
     "noImplicitAny": false,
-    "types": ["vite/client"],
+    "types": ["vite/client", "@testing-library/jest-dom"],
     "paths": {
       "@/*": ["./src/*"],
       "@sage-bionetworks/synapse-types": ["../../packages/synapse-types/src"],
