+++ packages/synapse-react-client/src/components/EntityFinder/tree/EntityTree.test.tsx
@@ -1,4 +1,3 @@
-import '@testing-library/jest-dom'
 import {
   EntityHeader,
   EntityPath,
