+++ packages/synapse-react-client/src/components/GoalsV2/GoalsV2.test.tsx
@@ -5,8 +5,7 @@ import {
   QueryResultBundle,
 } from '@sage-bionetworks/synapse-types'
 import { render, screen, waitFor } from '@testing-library/react'
-import '@testing-library/jest-dom'
-import { SynapseClient } from '../../index'
+import * as SynapseClient from '@/synapse-client/SynapseClient'
 import GoalsV2 from './GoalsV2'
 
 const tableQueryResult: QueryResultBundle = {
