+++ packages/synapse-react-client/tsconfig.json
@@ -6,7 +6,7 @@
     "allowJs": true,
     "isolatedModules": true,
     "noFallthroughCasesInSwitch": true,
-    "types": ["vitest/globals"],
+    "types": ["vitest/globals", "@testing-library/jest-dom"],
     "paths": {
       "@/*": ["./src/*"],
       // HMR will work if we uncomment these, but doing so causes various build/typecheck errors
